Match y_misc FUN_0042a0d0

+++ src/itfPanel.c
@@ -4336,7 +4336,7 @@ u64 FUN_003ada00(u64 param_1,int param_2)
 
   char msg[8];
 
-  u8 bVar1;
+  u16 bVar1;
 
   u8 uVar2;
